Make this a proper python package for easy installation
