+++ src/odusseas/machinelearning.py
@@ -1,6 +1,5 @@
 import os
-from dataclasses import dataclass
-from typing import Dict, List, Tuple
+from typing import Dict
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -10,24 +9,7 @@ from sklearn.linear_model._base import LinearModel
 from sklearn.metrics import explained_variance_score, r2_score
 from sklearn.model_selection import train_test_split
 
-from spec_utils import Spectrum
-
-error_matrix = {
-    "photometry": {
-        "115000": (0.10, 65),
-        "110000": (0.10, 68),
-        "94600": (0.12, 77),
-        "75000": (0.13, 78),
-        "48000": (0.13, 80),
-    },
-    "interferometry": {
-        "115000": (0.11, 90),
-        "110000": (0.11, 92),
-        "94600": (0.12, 95),
-        "75000": (0.13, 97),
-        "48000": (0.13, 99),
-    },
-}
+from odusseas.utils import Measurement, Reference, Result
 
 
 def make_plot(x, y, extra_coords, title, xlabel, ylabel, figsize, fout):
@@ -64,103 +46,6 @@ def get_regression_model(name: str) -> LinearModel:
         case "multitaskelasticnet":
             return linear_model.MultiTaskElasticNet(alpha=1)
     raise ValueError(f"Regression '{name}' is not valid")
-
-
-@dataclass
-class Reference:
-    resolution: int
-    reference: str
-
-    def __post_init__(self) -> None:
-        self.data = []
-        self.header = None
-        with open(self.fname) as f:
-            for i, line in enumerate(f):
-                if i > 0:
-                    self.data.append(line.strip().split(","))
-                else:
-                    self.header = line.strip().split(",")
-        if not self.header:
-            raise ValueError("Header not found")
-        self.data = np.array(self.data)
-
-    @property
-    def fname(self) -> str:
-        return f"reference_data/res{self.resolution}_{self.reference}RefEWPar.csv"
-
-    def subset_with_wavelengths(
-        self, wavelengths: np.ndarray
-    ) -> Tuple[pd.DataFrame, pd.DataFrame]:
-        res = dict()
-        for row, col in zip(self.data.T, self.header):
-            if col in ("names", "FeH", "Teff"):
-                res[col] = row
-                continue
-            wavelength = np.round(np.float64(col), 3)
-            if not wavelength in wavelengths:
-                continue
-            res[wavelength] = row
-        df = pd.DataFrame(res)
-        df_x = df.drop(["names", "FeH", "Teff"], axis=1).astype(float)
-        df_y = df[["names", "FeH", "Teff"]]
-        df_y.loc[:, "FeH"] = df_y["FeH"].astype(float)
-        df_y.loc[:, "Teff"] = df_y["Teff"].astype(float)
-        return df_x, df_y
-
-
-@dataclass
-class Measurement:
-    fname: str
-    resolution: int
-
-    def __post_init__(self) -> None:
-        self.spectrum = Spectrum(self.fname, self.resolution)
-
-        self.newdf = pd.read_csv(
-            os.path.join("results", f"{self.spectrum.name}.dat"),
-            header=None,
-            names=["wavelength", "EW"],
-        )
-        self.newdf.dropna(axis=1, inplace=True)
-        self.newdf[self.newdf.EW > 0.00001]
-        self.newdf.wavelength = self.newdf.wavelength.round(3)
-
-
-@dataclass
-class Result:
-    resolution: str
-    reference: str
-    spectrum: str
-    params_sub_result: Dict[str, List[float]]
-
-    def __str__(self) -> str:
-        d = self.params_sub_result
-        wide_error_FeH = round(
-            (
-                (np.std(d["fehs"])) ** 2
-                + (error_matrix[self.reference][self.resolution][0]) ** 2
-            )
-            ** (1 / 2),
-            3,
-        )
-        wide_error_Teff = round(
-            (
-                (np.std(d["teffs"])) ** 2
-                + (error_matrix[self.reference][self.resolution][1]) ** 2
-            )
-            ** (1 / 2),
-            3,
-        )
-        feh_mean = round(np.mean(d["fehs"]), 3)
-        feh_std = round(np.std(d["fehs"]), 3)
-        feh_mae = round(np.mean(d["feh_maes"]), 3)
-        teff_mean = int(np.mean(d["teffs"]))
-        teff_std = int(np.std(d["teffs"]))
-        teff_mae = int(np.mean(d["teff_maes"]))
-        r2_mean = round(np.mean(d["r2scores"]), 3)
-        var_mean = round(np.mean(d["variances"]), 3)
-
-        return f"{self.spectrum} {feh_mean} {feh_std} {feh_mae} {wide_error_FeH} {teff_mean} {teff_std} {teff_mae} {wide_error_Teff} {r2_mean} {var_mean}\n"
 
 
 def ML(spectra: Dict[str, int], regression: str, reference: str) -> None:
