+++ src/odusseas/main.py
@@ -1,24 +1,9 @@
-from enum import Enum
 from typing import Annotated
 
 import typer
 
-import machinelearning
-import spec_utils
-
-
-class Reference(str, Enum):
-    photometry = "photometry"
-    interferometry = "interferometry"
-
-
-class Regression(str, Enum):
-    linear = "linear"
-    ridge = "ridge"
-    ridgecv = "ridgecv"
-    multitasklasso = "multitasklasso"
-    multitaskelasticnet = "multitaskelasticnet "
-
+from odusseas import machinelearning, spec_utils
+from odusseas.utils import ReferenceEnum, RegressionEnum
 
 app = typer.Typer()
 
@@ -27,18 +12,18 @@ app = typer.Typer()
 def main(
     input_spectra: str,
     reference: Annotated[
-        Reference,
+        ReferenceEnum,
         typer.Option(
             help="choose the reference scale: 'photometry' for 65 stars with Teff from Casagrande08 and [Fe/H] from Neves12, or 'interferometry' for 47 stars with Teff from Khata21 and Rabus19, and [Fe/H] from Neves12",
             case_sensitive=False,
         ),
-    ] = Reference.interferometry,
+    ] = ReferenceEnum.interferometry,
     regression: Annotated[
-        Regression,
+        RegressionEnum,
         typer.Option(
             help="choose the ML model. Recommended: ridge", case_sensitive=False
         ),
-    ] = Regression.ridge,
+    ] = RegressionEnum.ridge,
     rv_cor: Annotated[bool, typer.Option()] = True,
     verbose: Annotated[bool, typer.Option()] = False,
     skip_ew_measurements: Annotated[
